@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_libs_app/routes.dart';
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -81,7 +81,7 @@ class _MyHomePageState extends State<MyHomePage> {
           ),
         ),
         onPressed: (){
-          callback?.call();
+          callback.call();
         },
       ),
     );
